requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
greenlet==3.1.0
idna==3.8
numpy==2.0.2
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
schedule==1.2.2
six==1.16.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2